requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
idna==2.10
joblib==0.16.0
nltk==3.5
numpy==1.19.1
openspeechlib==0.0.2
praat-textgrids==1.3.1
regex==2020.7.14
requests==2.24.0
scipy==1.5.1
tqdm==4.47.0
urllib3==1.25.9
